Services, Controllers, DTO's, Repository's voor 90% klaargezet, CRUD nog afronden. Testen, Auth nog doen.

+++ src/main/java/nl/novi/nepplaats/dto/categorie/CategorieDto.java
@@ -1,7 +1,7 @@
 package nl.novi.nepplaats.dto.categorie;
 
 public class CategorieDto {
-    private Integer categorieId;
+    private Long categorieId;
     private String categorieNaam;
     private String categorieBeschrijving;
 
@@ -13,15 +13,15 @@ public class CategorieDto {
     }
 
     // Constructor
-    public CategorieDto(Integer categorieId, String categorieNaam, String categorieBeschrijving) {
+    public CategorieDto(Long categorieId, String categorieNaam, String categorieBeschrijving) {
         this.categorieId = categorieId;
         this.categorieNaam = categorieNaam;
         this.categorieBeschrijving = categorieBeschrijving;
     }
 
     // Getters + Setters
-    public Integer getCategorieId() { return categorieId; }
-    public void setCategorieId(Integer categorieId) { this.categorieId = categorieId; }
+    public Long getCategorieId() { return categorieId; }
+    public void setCategorieId(Long categorieId) { this.categorieId = categorieId; }
 
     public String getCategorieNaam() { return categorieNaam; }
     public void setCategorieNaam(String categorieNaam) { this.categorieNaam = categorieNaam; }
